company_research_tool/requirements.txt
flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
celery==5.3.6
redis==5.0.1 